Move filterForkedElements to its own module.
