+++ api/get_map.js
@@ -3,8 +3,7 @@ var pumpify = require('pumpify')
 var mapStream = require('through2-map')
 var fromArray = require('from2-array')
 var collectTransform = require('../lib/collect-transform-stream')
-
-var cmpFork = require('../lib/util').cmpFork
+var filterForkedElements = require('../lib/filter_forked_elements')
 
 var refs2nodes = require('../lib/util').refs2nodes
 var checkRefExist = require('../lib/check_ref_ex.js')
@@ -44,67 +43,4 @@ module.exports = function (osm) {
 
 function filterForkedElementsStream () {
   return collectTransform(filterForkedElements)
-}
-
-function filterForkedElements (elements) {
-  var latestFirst = elements.sort(cmpFork)
-  var nonForkedElements = []
-  var keepNodeRefs = {}
-  var excludeNodeRefs = {}
-  var seen = {}
-
-  // Filter out the non-latest version of each element.
-  nonForkedElements = latestFirst.filter(function (element) {
-    if (!seen[element.id]) {
-      seen[element.id] = true
-
-      // Note that all of the nodes referenced by this way should be kept.
-      if (element.type === 'way') {
-        element.nodes.forEach(function (ref) {
-          keepNodeRefs[ref] = true
-        })
-      }
-
-      return true
-    } else {
-      seen[element.id] = true
-
-      // Note that all of the nodes referenced by this way should be culled.
-      if (element.type === 'way') {
-        element.nodes.forEach(function (ref) {
-          excludeNodeRefs[ref] = true
-        })
-      }
-
-      return false
-    }
-  })
-
-  // Remove excluded entries that appear in the keep entries.
-  Object.keys(keepNodeRefs).forEach(function (ref) {
-    if (excludeNodeRefs[ref]) {
-      delete excludeNodeRefs[ref]
-    }
-  })
-
-  // Filter out all nodes that are referenced in filtered ways.
-  nonForkedElements = nonForkedElements.filter(function (elm) {
-    if (elm.type === 'node' && (keepNodeRefs[elm.id] || !excludeNodeRefs[elm.id])) {
-      return true
-    } else if (elm.type !== 'node') {
-      return true
-    } else {
-      return false
-    }
-  })
-
-  // Sort by type
-  nonForkedElements.sort(cmpType)
-
-  return nonForkedElements
-}
-
-var typeOrder = { node: 0, way: 1, relation: 2 }
-function cmpType (a, b) {
-  return typeOrder[a.type] - typeOrder[b.type]
 }
